Refactored CliCommandWithGenericHost into GenericHostCliCommand
